Copy migration handling logic into a new crate for modularity
This crate is a result of building a system for handling database
migrations from scratch. It's become large enough that it's worth moving
out of my original project's codebase and into its own crate, which also
has the benefit that it can be shared on crates.io where others might
find it allows them to avoid building such a system from scratch like I
had to.